crewai_app/requirements.txt
crewai>=0.86.0
langchain>=0.3.0
langchain-ollama>=0.2.0
litellm>=1.50.0
requests>=2.32.0
python-dotenv>=1.0.0
streamlit>=1.40.0
mysql-connector-python>=9.0.0
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
